Moved item stack renderer to subpackage
Because more renderers are showing up on horizon!
